display tag signature
