+++ scm-ui/ui-types/src/Changesets.ts
@@ -26,6 +26,7 @@ import {Collection, Link, Links} from "./hal";
 import { Tag } from "./Tags";
 import { Branch } from "./Branches";
 import { Person } from "./Person";
+import {Signature} from "./Signature";
 
 export type Changeset = Collection & {
   id: string;
@@ -42,16 +43,7 @@ export type Changeset = Collection & {
   };
 };
 
-export type Signature = {
-  keyId: string;
-  type: string;
-  status: "VERIFIED" | "NOT_FOUND" | "INVALID";
-  owner?: string;
-  contacts?: Person[];
-  _links?: {
-    rawKey?: Link;
-  };
-}
+
 
 export type Contributor = {
   person: Person;
